Move BlogPostCard into Blog component folder
