+++ pages/home.js
@@ -13,7 +13,7 @@ import Banner from '../components/Banner'
 import Carousel from '../components/Carousel'
 import portraitPropTypes from '../propTypes/portrait'
 import Stat from '../components/Stat'
-import BlogPostCard from '../components/BlogPostCard'
+import BlogPostCard from '../components/Blog/BlogPostCard'
 import EmbeddedVideo from '../components/EmbeddedVideo'
 import PageSection from '../components/PageSection'
 import StatList from '../components/StatList'
